Make IdEqualityMixin shareable with other models
So we can use it on models that are nothing to do with broadcasts, it
needs to be moved to the place where our other generic/abstract models
live.
