+++ app/models/__init__.py
@@ -7,6 +7,18 @@ from notifications_utils.serialised_model import (
 )
 
 
+class IdEqualityMixin:
+
+    def __repr__(self):
+        return f'{self.__class__.__name__}(<{self.id}>)'
+
+    def __eq__(self, other):
+        return self.id == other.id
+
+    def __hash__(self):
+        return hash(self.id)
+
+
 class JSONModel(SerialisedModel):
 
     def __init__(self, _dict):
